refactor(entity): sistema de pontuação.

+++ gincana/gincana-entity/src/main/java/ifsul/gincana/entity/entities/Nota.java
@@ -1,7 +1,6 @@
 package ifsul.gincana.entity.entities;
 
 import java.io.Serializable;
-import java.util.Date;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
@@ -22,7 +21,7 @@ public class Nota implements Serializable {
     @NotNull
     @Basic(optional = false)
     @Column(name = "valor_nota")
-    private Double valor;
+    private int valor;
 
     @ManyToOne(cascade = ALL)
     @JoinColumn(name = "id_equipe")
@@ -40,11 +39,11 @@ public class Nota implements Serializable {
         this.id = id;
     }
 
-    public Double getValor() {
+    public int getValor() {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(int valor) {
         this.valor = valor;
     }
